estimated_histogram to support sampling
Taking time messurment of an operation can cause peformence degredation,
this patch adds sampling support for the estimated histogram.

It allows to add a sample with a counter that holds what is the actual
total number so far. So the samplling will be counted as multiple
entries in the estimated histogram.

The totatl count of the entries in the histogram will equal the _count
parameter.

Signed-off-by: Amnon Heiman <amnon@cloudius-systems.com>