storage_service: true snapshot size
For CFStats, one of the things needed is the size used by the snapshots. Since
the bulk of the work is map-reducing it and adding them together, we will just
call get_snapshot_details for the column family, and just selectively add just
what we need. No need for a separate method here.

Signed-off-by: Glauber Costa <glommer@scylladb.com>