Revert "do not try to write same memtable to sstable twice if moving it to a cache failed."
This reverts commit fff37d15cd02f1a5a864ac73140f53de6be006f3.

Says Tomek (and the comment in the code):

"update_cache() must be called before unlinking the memtable because cache + memtable at any time is supposed to be authoritative source of data for contained partitions. If there is a cache hit in cache, sstables won't be checked. If we unlink the memtable before cache is updated, it's possible that a query will miss data which was in that unlinked memtable, if it hits in the cache (with an old value)."